+++ packages/orq-rc/src/models/operations/updateconversation.ts
@@ -316,7 +316,7 @@ export const UpdateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01ketyde9199t3g0fhvfc5p8jt"),
+  _id: z.string().default("conv_01kev8pwe1ebmpdjb36x4mc47t"),
   kind: UpdateConversationKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),
@@ -348,7 +348,7 @@ export const UpdateConversationResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   UpdateConversationResponseBody
 > = z.object({
-  id: z.string().default("conv_01ketyde9199t3g0fhvfc5p8jt"),
+  id: z.string().default("conv_01kev8pwe1ebmpdjb36x4mc47t"),
   kind: UpdateConversationKind$outboundSchema,
   displayName: z.string(),
   createdAt: z.number(),