+++ packages/orq-rc/RELEASES.md
@@ -5769,3 +5769,13 @@ Based on:
 - [typescript v4.2.0-rc.23] packages/orq-rc
 ### Releases
 - [NPM v4.2.0-rc.23] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.23 - packages/orq-rc
+
+## 2026-01-13 08:51:37
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.685.0 (2.794.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.2.0-rc.24] packages/orq-rc
+### Releases
+- [NPM v4.2.0-rc.24] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.24 - packages/orq-rc